varios

+++ app/src/routes/api/admin/products/index.js
@@ -2,16 +2,19 @@ import db from '$lib/db';
 
 export async function post({ request, locals }) {
 	try {
-		const { producto } = await request.json();
-	
-		db.query(
-			`
+		if (local.user) {
+			const { producto } = await request.json();
+
+			db.query(
+				`
 		INSERT INTO "product"("name","brand","desc","price")
 		SELECT $1,$2,$3,$4
 		`,
-			[producto.nombre, producto.marca, producto.descripcion, producto.precio]
-		);
-		return { body: { hello: 'world' } };
+				[producto.nombre, producto.marca, producto.descripcion, producto.precio]
+			);
+			return { body: { hello: 'world' } };
+		}
+		return { body: {} };
 	} catch (error) {
 		return { body: { hello: 'world' } };
 	}
@@ -38,17 +41,20 @@ export async function put({ locals, url, request }) {
 	};
 }
 export async function del({ locals, url }) {
-	let product_id = url.searchParams.get('product_id');
+	if (locals.user) {
+		let product_id = url.searchParams.get('product_id');
 
-	await db.query(
-		`
+		await db.query(
+			`
 			DELETE FROM "product" WHERE product_id = $1
 	`,
-		[product_id]
-	);
-	return {
-		body: {
-			product: 'asd'
-		}
-	};
+			[product_id]
+		);
+		return {
+			body: {
+				product: 'asd'
+			}
+		};
+	}
+	return { body: {} };
 }
